Make the wiki downloader work with both versions of wiki
mushpedia uses the older layout.
In commit 0c18e061, I migrated away from table.mw-allpages-table-chunk
and to: ul.mw-allpages-chunk; but it turns out that was just a
versioning difference in the wiki.

It was hard to get it to work with both, so here's a fork that just
checks each.